+++ soya/src/data/redux/ReduxStore.js
@@ -8,6 +8,7 @@ import Hydration from './Hydration.js';
 
 import QueryResult from './QueryResult.js';
 import { compose, createStore, applyMiddleware } from 'redux';
+import { ActionTypes } from 'redux/lib/createStore';
 import thunk from 'redux-thunk';
 import scope from 'soya/lib/scope';
 
@@ -577,6 +578,7 @@ export default class ReduxStore extends Store {
       service = this.getService(serviceClass);
       this._serviceDependencies[id][serviceId] = service;
     }
+    this._store.dispatch({ type: ActionTypes.INIT });
   }
 
   /**